refacto backend computations in smart_plotter
